Fix custom `ease` & `interp` examples
Renamed `bevy_motiongfx_examples` to `bevy_examples`
