+++ examples/bevy_examples/examples/hello_world.rs
@@ -1,9 +1,9 @@
 use bevy::color::palettes;
 use bevy::pbr::NotShadowCaster;
 use bevy::prelude::*;
+use bevy_examples::timeline_movement;
 use bevy_motiongfx::prelude::*;
 use bevy_motiongfx::BevyMotionGfxPlugin;
-use bevy_motiongfx_examples::timeline_movement;
 
 fn main() {
     App::new()
